requirements.txt
# QuantBotForex - AI-Driven Forex Trading Platform
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing & Analysis
pandas>=2.2.0
numpy>=1.26.0
pandas-ta==0.3.14b0
ta==0.10.2
finta==1.3
TA-Lib==0.4.28

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1

# Trading & Financial
oandapyV20==0.7.2
ccxt==4.1.77

# Machine Learning
scikit-learn==1.3.2
tensorflow>=2.16.0
keras>=3.0.0

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
passlib==1.7.4

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# WebSocket Support
websockets==12.0

# Configuration
pyyaml==6.0.1
toml==0.10.2 